stress-file-ioctl: cast ~0 to uint16_t
Clean up implicit cast build warning

+++ stress-file-ioctl.c
@@ -280,7 +280,7 @@ static int stress_file_ioctl(const stress_args_t *args)
 			 * and exercise illegal dest_count to force an
 			 * ENOMEM error
 			 */
-			d->dest_count = ~0;
+			d->dest_count = (uint16_t)~0U;
 			ret = ioctl(fd, FIDEDUPERANGE, d);
 			(void)ret;
